+++ src/components/Page/index.js
@@ -1,11 +1,12 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { View, ViewPropTypes } from 'react-native';
-import { StatusBar } from 'react-native-simple-components';
 
 import styleConstants from '../../styleConstants';
 
 import styles from './styles';
+
+import StatusBar from '../StatusBar';
 
 const propTypes = {
   verticalCenter: PropTypes.bool,